old/plotmap2png/requirements.txt
pypng==0.0.19
